Improve create-stack build time by using OCI layout
The ihop.Client.Build method will now export the image from the Docker
daemon as soon as the build completes, writing it to disk in OCI layout
format. From this point on, the create-stack codebase will interact
exclusively with the OCI layout image format. The advantage here is that
we won't be shuttling an image back and forth between the Docker daemon
and program memory or disk for every operation we need to perform.
Instead, we can just manipulate the image on the local filesystem
directly using the GGCR layout package.