Add support for platform specific arguments
This commit adds support for platform specific arguments in the stack.toml file for use when creating multi-architecture stacks.

The Build and Run settings can contain arguments, i.e. `[build.args]`. The same set of arguments is applied across all of the platforms. However, in some cases it is necessary to have platform specific arguments. To do this you can define `[build.platforms.<platform>]` and supply a list of arguments. The platform-specific arguments are overlaid on the general arguments, so you can override general args with platform specific ones or add more arguments.

Signed-off-by: Daniel Mikusa <dan@mikusa.com>