Add an option to publish stack images directly
Instead of outputing OCI image files, this commit adds a `--publish` flag which signals `jam` to instead upload the created images directly to a registry. It will upload to the image reference listed in the `--buildOutput` and `--runOutput` arguments (instead of a path, use an image reference like `docker.io/foo/bar:0.0.2`.

Internally, this splits the client.Export code into two parts. One part that outputs the images to a temporary directory and antoher that either converts to an OCI file or uploads directly to the registry.

Signed-off-by: Daniel Mikusa <dan@mikusa.com>