inspector: support flattened buildpacks
With platforms introducing the ability to create flattened
buildpackages[1], the idea of 1 layer = 1 buildpack is not always true
anymore - a single layer can contain multiple buildpacks.
This change adds to the inspector the ability to read buildpack
metadata off flattened buildpacks.

This fixes "jam summarize" when creating summary/release notes
for flattened buildpacks. Previously it picked up the first
buildpack.toml it could find in the layer, and assumed it to represent
the buildpackage.

[1]: https://github.com/buildpacks/pack/pull/1691