update(uri): support for .cnb files in dependencies
Adds support for .cnb files for dependencies, leaving them untouched
during an update process. Supports both local paths, but an archive
from a http/https server alike.

I extracted the code from commands into its own file, to be able to
add a bit of test coverage.

Also extracted some of the check code (for suffix/prefix) into the uri
file to be able to reuse it and make it a bit more readable (hopefully).

LBNL, update the parser to keep http/https scheme around so the uri in
the dependencies doesn't get the scheme stripped.

Resolves: #358