feat(cnb): support for .cnb files and directories in dependencies
Adds support for .cnb files and directories for dependencies, leaving
them untouched during the update process. Supports cnb files from a
local paths and from a http/https server alike.

Extracted some of the check code (for suffix/prefix) into the uri
file to be able to reuse it and make it a bit more readable (hopefully).

LBNL, update the parser to keep http/https scheme around so the uri in
the dependencies doesn't get the scheme stripped.

Resolves: #358